Resetting of csv enrichments working
