+++ tests/data/csv/acoustic_speaker_enrichment.txt
@@ -1,0 +1,2 @@
+speaker,property_one
+unknown,something
